Added a new feature "--get-summaries" to retrieve Summary Activity Tracker information
This new flag "--get-summaries" fetches the Summary Activity Tracker
information from the watch. These summaries are stored in the files 0x00b3xxxx.
These files use Google's protobuffer format, and hence need to be parsed by a
prototobuf library. The one that is used here is available as a package on
Debian and Ubuntu.

The summaries information can be exported as CSV file too, by using
--get-summaries --set-formats=csv. I am not intending to write a tool akin to
ttbincnv at this moment.

This feature can be extended to also parse 0x00b1xxxx files, but that is a
future step.